requirements.txt
streamlit
flask
flask_cors
pdfplumber
langchain
langchain-google-genai
faiss-cpu
google-generativeai
python-dotenv
langchain-community
PyPDF2
gunicorn
